import mongo-c-driver 1.16.2

+++ src/mongoc/mongoc-write-concern-private.h
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-#include "mongoc/mongoc-prelude.h"
+#include "mongoc-prelude.h"
 
 #ifndef MONGOC_WRITE_CONCERN_PRIVATE_H
 #define MONGOC_WRITE_CONCERN_PRIVATE_H
@@ -33,7 +33,7 @@ struct _mongoc_write_concern_t {
    int8_t fsync_; /* deprecated */
    int8_t journal;
    int32_t w;
-   int32_t wtimeout;
+   int64_t wtimeout;
    char *wtag;
    bool frozen;
    bson_t compiled;
